refactor(agentic): decompose functions and simplify branching across tools and test suites
